# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
...
    "/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_delete",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "FinStack API",
    "description": "REST API for FinStack expense, policy, audit, reporting, and dashboard workflows.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "RBAC role: superuser, admin, or user"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "expense_submitter",
                "manager",
                "finance_officer",
                "compliance_officer",
                "configuration_manager"
              ]
            }
          },
          "managerEmployeeId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive"
            ]
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "approved",
              "pending",
              "rejected"
            ]
          },
          "organizationId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstLoginRequired": {
            "type": "boolean"
          }
        },
        "required": [
          "employeeId",
          "fullName",
          "email",
          "roles",
          "organizationId"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "expense_submitter",
                "manager",
                "finance_officer",
                "compliance_officer",
                "configuration_manager"
              ]
            }
          },
          "managerEmployeeId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive"
            ]
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "approved",
              "pending",
              "rejected"
            ]
          },
          "organizationId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstLoginRequired": {
            "type": "boolean"
          }
        }
      },
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "expenseId": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "merchant": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processed",
              "failed",
              "reconciled"
            ]
          },
          "transactionDate": {
            "type": "string"
          },
          "processedAt": {
            "type": "string"
          }
        },
        "required": [
          "expenseId",
          "employeeId",
          "organizationId",
          "amount",
          "merchant",
          "categoryId",
          "paymentMethod",
          "transactionDate"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "expenseId": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "merchant": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processed",
              "failed",
              "reconciled"
            ]
          },
          "transactionDate": {
            "type": "string"
          },
          "processedAt": {
            "type": "string"
          }
        }
      }
    }
  }
}